\name{ropenphactslibs-package}
\alias{ropenphactslibs-package}
\alias{ropenphactslibs}
\docType{package}
\title{
Apache and ops4j libraries to work with RDF data.
}
\description{
Package with the ops4j library jars, mostly aimed for use
in the ropenphacts package.
}
\details{
\tabular{ll}{
Package: \tab ropenphactslibs\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2013-12-21\cr
License: \tab AGPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Egon Willighagen

Maintainer: Egon Willighagen <egon.willighagen@gmail.com>
}
\keyword{ package }
\keyword{ Open PHACTS }
\keyword{ RDF }
